requirements.txt
google-auth
google-cloud-pubsub
mailtrap